[Customoverlayの状況]を態々intに変換していた問題を修正

+++ SuperNewRoles/Modules/CustomOptionModel.cs
@@ -1159,7 +1159,7 @@ public static class GameOptionsNextPagePatch
             if (AmongUsClient.Instance.GameState != InnerNet.InnerNetClient.GameStates.Started)
                 SuperNewRolesPlugin.optionsPage++;
             // 試合中はRegulationのoverlayを表示している時のみ, 2ページ単位でページを送る
-            else if (CustomOverlays.nowPattern == (int)CustomOverlays.CustomOverlayPattern.Regulation) SuperNewRolesPlugin.optionsPage += 2;
+            else if (CustomOverlays.nowPattern == CustomOverlays.CustomOverlayPattern.Regulation) SuperNewRolesPlugin.optionsPage += 2;
 
             // ページが最大ページを超えたら, ページを0に戻す
             if (SuperNewRolesPlugin.optionsPage > SuperNewRolesPlugin.optionsMaxPage)
